mmc: extract FileBufferedWriter
